requirements.txt
blinker==1.9.0
click==8.2.1
colorama==0.4.6
Flask==2.3.3
gmpy2==2.2.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
mpmath==1.3.0
pycryptodome==3.23.0
sympy==1.14.0
Werkzeug==2.3.7
